+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/BottomFooter.js
