+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/Cloud.js
