+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/Copyright.js
