+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/DivSixImage.js
