+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/LeftCloud.js
