+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/PhundLogo.js
