+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/RSTriangleSVG.js
