+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/RedBorrowButton.js
@@ -1,36 +1,34 @@
-import {Button, Row, Col} from 'antd'
 import React from 'react'
+import {Button, Row, Col} from 'antd'
 import {RightCircleOutlined} from '@ant-design/icons'
-import '../Landing/CSS/Button.css'
+import '../../../Landing/CSS/Button.css'
 
+function ClickBorrow() {
+    const Buttonstyles = {
+    width: "508px",
+    height: "135px",
+    marginTop:"10",
+    paddingTop: "15",
+    paddingBottom: "15",
+    marginLeft: "30",
+    marginRight: "30",
+    backgroundColor:'#FF5776',
+    borderRadius: "25px",
+    borderWidth: "1",
 
+}
 
-
-
-function ClickLend() {
-  
     return(
-        <>
-          <Button
-          className='blue'
-          style={{
-              width: "508px",
-              height: "135px",
-              marginTop:"10",
-              paddingTop: "15",
-              paddingBottom: "15",
-              marginLeft: "30",
-              marginRight: "30",
-              backgroundColor:'#71B8F2',
-              borderRadius: "25px",
-              borderWidth: "1",
-                }}
-          >
-            <Row>
+        <Button
+        style={Buttonstyles}
+        className="red"
+        >
+
+<Row>
               <p
               style={{alignItems: 'flex-start', fontSize: 30, position: 'relative', marginTop: 10, fontWeight: 'bold', marginBottom: 0}}
               >
-              Lend.
+              Borrow.
               </p>
             </Row>
             <Row>
@@ -40,7 +38,7 @@ function ClickLend() {
                 <p
                 style={{ fontSize: 20 }}
                 >
-                Investment Porfolios starting at $25
+                Insant loans up to $500
               </p>
               </Col>
               <Col
@@ -54,10 +52,10 @@ function ClickLend() {
                 style={{color: "#FFFFFF", fontSize: 50, marginBottom: 15}} 
               />
               </Col>
-              </Row>       
-          </Button>
-        </>
-)
+              </Row>
+
+        </Button>
+        )
 }
 
-export default ClickLend
+export default ClickBorrow
