+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/SmallBlueLendBtn.js
