+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/SmallRedBorrowButton.js
