+++ client/src/components/Borrow/utils/Landing/Assets/DivSixComponents/triangleIcon.js
