+++ client/src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 
-import { store } from './helpers';
+import { store } from '../Redux/helpers';
 import { App } from './App.jsx';
